Add fallback service

+++ core/src/main/java/dev/nick/imageloader/display/ImageSettableIdCreator.java
@@ -19,5 +19,5 @@ package dev.nick.imageloader.display;
 import android.support.annotation.NonNull;
 
 public interface ImageSettableIdCreator {
-    int createSettableId(@NonNull ImageSettable settable);
+    long createSettableId(@NonNull ImageSettable settable);
 }
